rest-service/requirements.txt
cassandra-driver==3.25.0
click==8.1.3
Flask==2.1.2
geomet==0.2.1.post1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
six==1.16.0
Werkzeug==2.1.2
